+++ Assets/Enemies/Scripts/Enemy.cs
@@ -11,7 +11,7 @@ using BehaviourTree;
 [RequireComponent(typeof(Rigidbody2D))]
 public abstract class Enemy : BehaviourTree.Tree
 {
-    [SerializeField] float playerCollisionDamage = 10; //The damage that will be dealt to the player if they walk into the enemy
+    [SerializeField] int playerCollisionDamage = 10; //The damage that will be dealt to the player if they walk into the enemy
     [SerializeField] Vector2 playerCollisionForce = new Vector2(2000, 0.5f); //The force that will be added to the player if they walk into the enemy
     protected EnemyStats stats;
     protected Animator animator;
